+++ lesson18/src/main/java/ru/otus/igorr/books/lesson18/config/MongoProps.java
@@ -11,6 +11,7 @@ import org.springframework.stereotype.Component;
 @Component
 @ConfigurationProperties("spring.data.mongodb")
 public class MongoProps {
+    private String host;
     private int port;
     private String database;
     private String uri;
