--- lesson18/src/main/java/ru/otus/igorr/books/lesson18/controllers/ResponseId.java
@@ -1,14 +1,0 @@
-package ru.otus.igorr.books.lesson18.controllers;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ResponseId {
-    private String id;
-
-    public ResponseId(String id) {
-        this.id = id;
-    }
-}
