--- lesson18/src/main/java/ru/otus/igorr/books/lesson18/service/author/AuthorReactiveServiceImpl.java
@@ -1,53 +1,0 @@
-package ru.otus.igorr.books.lesson18.service.author;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.stereotype.Service;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
-import ru.otus.igorr.books.lesson18.domain.author.Author;
-import ru.otus.igorr.books.lesson18.dto.AuthorDto;
-import ru.otus.igorr.books.lesson18.dto.DtoConverter;
-import ru.otus.igorr.books.lesson18.repository.author.AuthorReactiveRepository;
-
-@Service
-public class AuthorReactiveServiceImpl implements AuthorReactiveService {
-
-    private final AuthorReactiveRepository repository;
-    private final DtoConverter<Author, AuthorDto> converter;
-
-    @Autowired
-    public AuthorReactiveServiceImpl(AuthorReactiveRepository repository,
-                                     @Qualifier("authorConverter") DtoConverter converter) {
-        this.repository = repository;
-        this.converter = converter;
-    }
-
-    @Override
-    public Mono<AuthorDto> get(String id) {
-        return repository.findById(id).map(author -> converter.convert(author));
-    }
-
-    @Override
-    public Mono<AuthorDto> add(AuthorDto dto) {
-        return repository.save(converter.fill(dto))
-                .map(author -> converter.convert(author));
-    }
-
-    @Override
-    public Flux<AuthorDto> getList() {
-        return repository.findAll()
-                .map(author -> converter.convert(author));
-    }
-
-    @Override
-    public Flux<AuthorDto> getListByName(String mask) {
-        // TODO: 31.05.2019
-        return null;
-    }
-
-    @Override
-    public Mono<Void> delete(String id) {
-        return repository.deleteById(id);
-    }
-}
