--- lesson18/src/main/java/ru/otus/igorr/books/lesson18/service/facade/RServicesFacade.java
@@ -1,52 +1,0 @@
-package ru.otus.igorr.books.lesson18.service.facade;
-
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
-import ru.otus.igorr.books.lesson18.dto.AuthorDto;
-import ru.otus.igorr.books.lesson18.dto.BookDto;
-import ru.otus.igorr.books.lesson18.dto.GenreDto;
-import ru.otus.igorr.books.lesson18.dto.NoteDto;
-
-public interface RServicesFacade {
-
-    // Genre
-    Flux<GenreDto> getGenreList();
-
-    Mono<GenreDto> getGenre(String id);
-
-    Mono<GenreDto> addGenre(GenreDto genre);
-
-    Flux<GenreDto> getGenreListByName(String mask);
-
-    Mono<Void> deleteGenre(String id);
-
-    Mono<Void> deleteGenre(GenreDto genreDto);
-
-
-    // Author services
-    Mono<AuthorDto> getAuthor(String id);
-
-    Flux<AuthorDto> getAuthorList();
-
-    Flux<AuthorDto> getAuthorListByName(String mask);
-
-    Mono<AuthorDto> addAuthor(AuthorDto authorDto);
-
-    Mono<Void> deleteAuthor(String id);
-
-
-    // Book services
-    Flux<BookDto> getBookList();
-
-    Mono<BookDto> getBook(String id);
-
-    Mono<BookDto> addBook(BookDto bookDto);
-
-    Mono<Void> deleteBook(String id);
-
-    Mono<NoteDto> addNote(NoteDto noteDto);
-
-    Mono<Void> deleteNote(String noteId);
-
-
-}
