+++ lesson18/src/main/resources/application.yaml
@@ -13,6 +13,8 @@ spring:
   #
   data:
     mongodb:
+      host: localhost
+      port: 27017
       database: testdb
 
   batch:
