+++ lesson18/src/test/java/ru/otus/igorr/books/lesson18/handlers/GenreHandler.java
@@ -12,7 +12,6 @@ import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.web.reactive.function.server.ServerResponse.ok;
 
 
-
 public class GenreHandler {
 
     private final GenreController controller;
@@ -24,34 +23,17 @@ public class GenreHandler {
     public Mono<ServerResponse> listPage(ServerRequest request) {
         Flux<GenreDto> flux = controller.listPageFlux();
         return ok().contentType(APPLICATION_JSON).body(flux, GenreDto.class);
-        //return null;
     }
 
     public Mono<ServerResponse> viewPage(ServerRequest request) {
         Mono<GenreDto> mono = controller.viewPageFlux(request.pathVariable("id"));
         return ok().contentType(APPLICATION_JSON).body(mono, GenreDto.class);
-        //return null;
     }
 
     public Mono<ServerResponse> addPage(ServerRequest request) {
         Mono<GenreDto> mono = controller.addPageFlux(request.bodyToMono(GenreDto.class).block());
         return ok().contentType(APPLICATION_JSON).body(mono, GenreDto.class);
     }
-
-
-    /*
-    public Mono<ServerResponse> viewPage(ServerRequest request) {
-        Mono<Genre> genre = request.bodyToMono(Person.class);
-        return ok().build(repository.savePerson(genre));
-    }
-
-    public Mono<ServerResponse> getPerson(ServerRequest request) {
-        int personId = Integer.valueOf(request.pathVariable("id"));
-        return repository.getPerson(personId)
-                .flatMap(person -> ok().contentType(APPLICATION_JSON).body(fromObject(person)))
-                .switchIfEmpty(ServerResponse.notFound().build());
-    }
-    */
 
 
 }
